+++ Modules/SharePointDsc/DSCResources/MSFT_SPAlternateUrl/MSFT_SPAlternateUrl.psm1
@@ -6,16 +6,20 @@ function Get-TargetResource
     (
         [Parameter(Mandatory = $true)]
         [System.String] 
-        $WebAppUrl,
+        $WebAppName,
 
         [Parameter(Mandatory = $true)]  
         [ValidateSet("Default","Intranet","Extranet","Custom","Internet")] 
         [System.String] 
         $Zone,
 
-        [Parameter(Mandatory = $false)] 
+        [Parameter(Mandatory = $true)] 
         [System.String] 
         $Url,
+
+        [Parameter(Mandatory = $false)] 
+        [System.Boolean] 
+        $Internal = $false,
 
         [Parameter(Mandatory = $false)] 
         [ValidateSet("Present","Absent")] 
@@ -27,40 +31,40 @@ function Get-TargetResource
         $InstallAccount
     )
 
-    Write-Verbose -Message "Getting Alternate URL for $Zone in $WebAppUrl"
+    Write-Verbose -Message "Getting Alternate URL for $Zone in $WebAppName"
 
     $result = Invoke-SPDSCCommand -Credential $InstallAccount `
                                   -Arguments $PSBoundParameters `
                                   -ScriptBlock {
         $params = $args[0]
 
-        $aam = Get-SPAlternateURL -WebApplication $params.WebAppUrl `
-                                  -Zone $params.Zone `
-                                  -ErrorAction SilentlyContinue | Select-Object -First 1
-        $url = $null
-        $Ensure = "Absent"
-        
-        if (($null -eq $aam) -and ($params.Zone -eq "Default")) 
+        $aam = Get-SPAlternateURL -Identity $params.Url `
+                                  -ErrorAction SilentlyContinue
+
+        if ($null -eq $aam) 
         {
-            # The default zone URL will change the URL of the web app, so assuming it has been applied
-            # correctly then the first call there will fail as the WebAppUrl parameter will no longer
-            # be the URL of the web app. So the assumption is that if a matching default entry with
-            # the new public URL is found then it can be returned and will pass a test.
-            $aam = Get-SPAlternateURL -Zone $params.Zone | Where-Object -FilterScript { 
-                $_.PublicUrl -eq $params.Url 
-            } | Select-Object -First 1
+            return @{
+                WebAppName = $params.WebAppName
+                Zone = $params.Zone
+                Url = $params.Url
+                Ensure = "Absent"
+            }
         }
-        
-        if ($null -ne $aam) {
-            $url = $aam.PublicUrl
-            $Ensure = "Present"
+
+        $internal = $false
+        if ($aam.PublicUrl -ne $aam.IncomingUrl)
+        {
+            $internal = $true
         }
-        
+
+        $wa = Get-SPWebApplication -Identity $aam.PublicUrl
+
         return @{
-            WebAppUrl = $params.WebAppUrl
-            Zone = $params.Zone
-            Url = $url
-            Ensure = $Ensure
+            WebAppName = $wa.Name
+            Zone = $aam.Zone
+            Url = $aam.IncomingUrl
+            Internal = $internal
+            Ensure = "Present"
             InstallAccount = $params.InstallAccount
         }
     }
@@ -74,17 +78,21 @@ function Set-TargetResource
     (
         [Parameter(Mandatory = $true)]
         [System.String] 
-        $WebAppUrl,
+        $WebAppName,
 
         [Parameter(Mandatory = $true)]  
         [ValidateSet("Default","Intranet","Extranet","Custom","Internet")] 
         [System.String] 
         $Zone,
 
-        [Parameter(Mandatory = $false)] 
+        [Parameter(Mandatory = $true)] 
         [System.String] 
         $Url,
 
+        [Parameter(Mandatory = $false)] 
+        [System.Boolean] 
+        $Internal = $false,
+        
         [Parameter(Mandatory = $false)] 
         [ValidateSet("Present","Absent")] 
         [System.String] 
@@ -96,33 +104,97 @@ function Set-TargetResource
 
     )
 
-    Write-Verbose -Message "Setting Alternate URL for $Zone in $WebAppUrl"
+    Write-Verbose -Message "Setting Alternate URL for $Zone in $WebAppName"
 
-    $CurrentValues = Get-TargetResource @PSBoundParameters
-    
     if ($Ensure -eq "Present") 
     {
-        if ([string]::IsNullOrEmpty($Url)) 
-        {
-            throw "URL must be specified when ensure is set to present"
-        }
-
         Invoke-SPDSCCommand -Credential $InstallAccount `
-                            -Arguments ($PSBoundParameters, $CurrentValues) `
+                            -Arguments $PSBoundParameters `
                             -ScriptBlock {
             $params = $args[0]
-            $CurrentValues = $args[1]
 
-            if ([string]::IsNullOrEmpty($CurrentValues.Url)) 
+            $webapp = Get-SPWebApplication $params.WebAppName
+
+            if ($null -eq $webapp)
             {
-                New-SPAlternateURL -WebApplication $params.WebAppUrl `
-                                   -Url $params.Url `
-                                   -Zone $params.Zone
+                throw "Web application was not found. Please check WebAppName parameter!"
+            }
+
+            $urlAam = Get-SPAlternateURL -Identity $params.Url `
+                                         -ErrorAction SilentlyContinue
+
+            $webAppAams = Get-SPAlternateURL -WebApplication $params.WebAppName `
+                                             -Zone $params.Zone `
+                                             -ErrorAction SilentlyContinue
+
+            if ($null -eq $webAppAams) 
+            {
+                # No AAM found on specified WebApp in specified Zone
+                if ($null -eq $urlAam)
+                {
+                    # urlAAM not found, so it is safe to create AAM on specified zone
+                    $cmdParams = @{
+                        WebApplication = $params.WebAppName
+                        Url = $params.Url
+                        Zone = $params.Zone
+                    }
+                    if (($params.ContainsKey("Internal") -eq $true))
+                    {
+                        $cmdParams.Add("Internal", $params.Internal)
+                    }
+                    New-SPAlternateURL @cmdParams | Out-Null
+                }
+                else
+                {
+                    throw ("Specified URL found on different WebApp/Zone: WebApp $($urlAam.PublicUrl) in zone $($urlAam.Zone)")
+                }
             } 
             else 
             {
-                Get-SPAlternateURL -WebApplication $params.WebAppUrl `
-                                   -Zone $params.Zone | Set-SPAlternateURL -Url $params.Url
+                # WebApp has one or more AAMs, check for URL
+                $aamForUrl = $webAppAams | Where-Object -FilterScript { $_.IncomingUrl -eq $params.Url }
+
+                if ($null -eq $aamForUrl)
+                {
+                    # URL not configured on WebApp
+                   if ($null -eq $urlAam)
+                    {
+                        # urlAAM not found, so it is safe to create AAM on specified zone
+                        $cmdParams = @{
+                            WebApplication = $params.WebAppName
+                            Url = $params.Url
+                            Zone = $params.Zone
+                        }
+                        if (($params.ContainsKey("Internal") -eq $true))
+                        {
+                            $cmdParams.Add("Internal", $params.Internal)
+                        }
+                        New-SPAlternateURL @cmdParams | Out-Null
+                    }
+                    else
+                    {
+                        throw ("Specified URL ($($params.Url)) found on different WebApp/Zone: WebApp $($urlAam.PublicUrl) in zone $($urlAam.Zone)")
+                    }
+                 }
+                 else
+                 {
+                    if ($params.Internal -eq $false)
+                    {
+                        if (($urlAam.PublicUrl -eq $aamForUrl.PublicUrl) -and `
+                            ($urlAam.Zone -eq $aamForUrl.Zone))
+                        {
+                            $webAppAams | Set-SPAlternateURL -Url $params.Url | Out-Null
+                        }
+                        else
+                        {
+                            throw ("Specified URL found on different WebApp/Zone: WebApp $($urlAam.PublicUrl) in zone $($urlAam.Zone)")
+                        }
+                    }
+                    else
+                    {
+                        Write-Verbose -Message "URL already exists!"
+                    }
+                 }
             }
         }
     } 
@@ -132,8 +204,10 @@ function Set-TargetResource
                             -Arguments $PSBoundParameters `
                             -ScriptBlock {
             $params = $args[0]
-            Get-SPAlternateURL -WebApplication $params.WebAppUrl `
-                               -Zone $params.Zone | Remove-SPAlternateURL -Confirm:$false
+            $aam = Get-SPAlternateURL -Identity $params.Url `
+                                      -ErrorAction SilentlyContinue
+
+            Remove-SPAlternateURL -Identity $aam -Confirm:$false
         }
     }    
 }
@@ -146,16 +220,20 @@ function Test-TargetResource
     (
         [Parameter(Mandatory = $true)]
         [System.String] 
-        $WebAppUrl,
+        $WebAppName,
 
         [Parameter(Mandatory = $true)]  
         [ValidateSet("Default","Intranet","Extranet","Custom","Internet")] 
         [System.String] 
         $Zone,
 
-        [Parameter(Mandatory = $false)] 
+        [Parameter(Mandatory = $true)] 
         [System.String] 
         $Url,
+
+        [Parameter(Mandatory = $false)] 
+        [System.Boolean] 
+        $Internal = $false,
 
         [Parameter(Mandatory = $false)] 
         [ValidateSet("Present","Absent")] 
@@ -168,20 +246,20 @@ function Test-TargetResource
 
     )
 
-    Write-Verbose -Message "Testing Alternate URL for $Zone in $WebAppUrl"
+    Write-Verbose -Message "Testing Alternate URL for $Zone in $WebAppName"
     
     $PSBoundParameters.Ensure = $Ensure
-    
-    if ([string]::IsNullOrEmpty($Url) -and $Ensure -eq "Present") 
-    {
-        throw "URL must be specified when ensure is set to present"
-    }
+    $PSBoundParameters.Internal = $Internal
 
     if ($Ensure -eq "Present")
     {
         return Test-SPDscParameterState -CurrentValues (Get-TargetResource @PSBoundParameters) `
                                         -DesiredValues $PSBoundParameters `
-                                        -ValuesToCheck @("Url", "Ensure")
+                                        -ValuesToCheck @("WebAppName", `
+                                                         "Zone", `
+                                                         "Url", `
+                                                         "Ensure", `
+                                                         "Internal")
     }
     else 
     {